Move utils and helpers to Mu-module
To avoid collisions with other packages, mu-utility functions will not
be included as globals anymore by default, but should be used with the
Mu-namespace (e.g. Mu.query instead of query).

Providing the INCLUDE_LEGACY_UTILS flag enables previous behaviour and
includes the legacy SinatraTemplate::Utils globally. This may smoothen
the transition to the new mu-ruby-template for existing
services. Deprecation messages will be printed when those legacy utils
are invoked.
