+++ sinatra_template/sparql/client.rb
@@ -1,26 +1,14 @@
-require 'sparql/client'
+require_relative '../../mu.rb'
 
 module SinatraTemplate
   module SPARQL
-    class Client < ::SPARQL::Client
-      ##
-      # Performs an HTTP request against the SPARQL endpoint.
-      #
-      # @param  [String, #to_s]          query
-      # @param  [Hash{String => String}] headers
-      # @yield  [response]
-      # @yieldparam [Net::HTTPResponse] response
-      # @return [Net::HTTPResponse]
-      # @see    http://www.w3.org/TR/sparql11-protocol/#query-operation
-      def request(query, headers = {}, &block)
-        headers['mu-session-id'] = RequestStore.store[:mu_session_id]
-        headers['mu-call-id'] = RequestStore.store[:mu_call_id]
-        headers['mu-auth-allowed-groups'] = RequestStore.store[:mu_auth_allowed_groups]
-        headers['mu-auth-used-groups'] = RequestStore.store[:mu_auth_used_groups]
-        response = super
-        RequestStore.store[:mu_auth_allowed_groups] = response['mu-auth-allowed-groups']
-        RequestStore.store[:mu_auth_used_groups] = response['mu-auth-used=groups']
-        response
+    class Client
+      # Forward all methods to Mu::SPARQL::Client
+      Mu::SPARQL::Client.instance_methods.each do |method|
+        define_method(method) do |*args, &block|
+          Mu.log.warn "[DEPRECATION] #{SinatraTemplate::SPARQL::Client.name}.#{__callee__} is deprecated. Please use #{Mu::SPARQL::Client.name}.#{__method__} instead."
+          Mu::SPARQL::Client.send(method, *args, &block)
+        end
       end
     end
   end
